app/requirements.txt
setuptools
pywebview
pyinstaller
numpy